requirements.txt
transformers
torch
accelerate
bitsandbytes
chromadb
sentence-transformers
gradio
rich
fastapi
uvicorn
python-dotenv
loguru
pypdf2
beautifulsoup4
faiss-cpu
numpy
pandas
scikit-learn
llama-cpp-python
psutil
GPUtil
huggingface-hub